Update install.sh to work on Alpine

+++ install.sh
@@ -3,7 +3,11 @@
 # Hack to make my dotfiles in Github Codespaces
 # Install stow
 if [ "$CODESPACES" == "true" ]; then
-    sudo apt-get update && sudo apt-get -y install --no-install-recommends stow
+    if [ -f /etc/alpine-release ]; then
+        sudo apk add stow
+    else
+        sudo apt-get update && sudo apt-get -y install --no-install-recommends stow
+    fi
     
     stow -t ~ git neovim
